+++ src/imio/smartweb/common/browser/static/webpack.config.js
@@ -9,6 +9,7 @@ module.exports = {
     entry: {
         view: './src/view.js',
         edit: './src/edit.js',
+        tinymce: './src/tinymce.js',
     },
     output: {
         filename: 'smartweb-common-[name]-compiled.js',